refactor: handlers/manifestoのリファクタリングとファイル名変更
- マニフェスト作成ロジックを独立した関数に分離
- ハンドラー関数を別関数として切り出し、責務を明確化
- ファイル名をmanifesto.tsからmanifesto_create.tsに変更
- 関数の責務に合わせた適切な命名に改善
